FDE PROJECT FILE/project/requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pandas==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9
influxdb-client==1.40.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
ta==0.11.0
scikit-learn==1.4.0
xgboost==2.0.3
prometheus-client==0.19.0
py-healthcheck==1.10.1
python-json-logger==2.0.7
websocket-client==1.7.0
pyarrow==14.0.2
pyyaml==6.0.1
supabase==2.3.4
asyncpg==0.29.0
